breaking: refactor of the cordova plugin
